==> 03_selenium_scrapping.ipynb <==
## Scrapping Autoplius.lt

# Importuojame  bibliotekas

import pandas as pd
import time
from selenium import webdriver
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.action_chains import ActionChains
from selenium.webdriver.common.alert import Alert

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)
pd.set_option('display.max_colwidth', None)

# Scrapiname parduodamų elektrinių automobilių duomeni iš autoplius.lt
url = 'https://m.autoplius.lt/skelbimai/naudoti-automobiliai/?fuel_id%5B35%5D=35'
driver = webdriver.Chrome()
driver.get(url)
time.sleep(20)

autoplius_modeliai = []
autoplius_kainos = []
autoplius_g_datos = []

df = pd.DataFrame()

# Naršyklės inspect pagalba gauti keliai
#//*[@id="container"]/div[5]/div[2]/a[2]/div[2]/div[1]/div[1]/div[1]      metai
#//*[@id="container"]/div[5]/div[2]/a[2]/div[2]/div[1]/div[1]/div[2]/div  modelis
#//*[@id="container"]/div[5]/div[2]/a[2]/div[2]/div[1]/div[2]/strong      kaina

i = 0
while True:

    ap_modeliai = driver.find_elements(By.XPATH, '//*[@id="container"]/div/div/a/div/div/div/div[2]/div')
    ap_g_datos = driver.find_elements(By.XPATH, '//*[@id="container"]/div/div/a/div/div/div[1]/div[1]')
    ap_kainos = driver.find_elements(By.XPATH, '//*[@id="container"]/div/div/a/div/div/div/strong')
   

    for kaina, modelis in zip(ap_kainos, ap_modeliai):
        # vietomis kaina su nuolaida
        autoplius_kainos.append(kaina.text.split('\n')[0].strip().replace(' ', '')[:-1])
        
        autoplius_modeliai.append(modelis.text.strip())
        
    for datos in ap_g_datos:
        if len(datos.text) != 0:
            autoplius_g_datos.append(datos.text.strip())




    print('nuskanuotas psl nr:', i)
    print('ilgiai', len(autoplius_modeliai), len(autoplius_kainos), len(autoplius_g_datos))
    if  (len(autoplius_modeliai) != len(autoplius_g_datos)) and (len(autoplius_kainos) != len(autoplius_g_datos)):
        break
    
    # time.sleep(30)
    # mygtukas ATGAL atsiranda tik po pirmo puslapio
    try: 
        if i == 0:
            all_matches_button = driver.find_element(By.XPATH, '//*[@id="container"]/div[5]/div[2]/div[6]/a')
        else:
            all_matches_button = driver.find_element(By.XPATH, '//*[@id="container"]/div[5]/div[2]/div[6]/a[2]')
        time.sleep(20)
        all_matches_button.click()
        
    except Exception as e:
        print(f"nerastas mygtukas Pirmyn arba nėra sekančio puslapio: {e}")
        print('Skenavimas baigtas')
        driver.quit()
        break

    i += 1

# Sudedame gautą informaciją į df
df = pd.DataFrame({'Modelis':autoplius_modeliai, 'Kaina': autoplius_kainos, 'Metai': autoplius_g_datos})

# Išsaugome duomenis faile
df.to_csv('autoplius_EV.csv', index=False)

print(df.head())

driver.quit()
