==> 06_main_visi_v2.ipynb <==
## Sukuriame pagrindinį darbinį failą su 3 stulpeliais

# importuojam bibliotekas
import numpy as np
import pandas as pd

# Nuskaitome duomenis iš failo
df_visi_v2 = pd.read_csv('data_visi.csv')

# Konvertuojame stulpelį į datetime formatą
df_visi_v2['PIRM_REG_DATA_LT_1'] = pd.to_datetime(df_visi_v2['PIRM_REG_DATA_LT'])
df_visi_v2

# Sukuriame du naujus stulpelius
df_visi_v2['PIRM_REG_METAI_LT'] = df_visi_v2['PIRM_REG_DATA_LT_1'].dt.strftime('%Y')
df_visi_v2['PIRM_REG_MEN_LT'] =   df_visi_v2['PIRM_REG_DATA_LT_1'].dt.strftime('%m')
df_visi_v2

# # Konvertuojame reikšmes į int
df_visi_v2['PIRM_REG_METAI_LT'] = df_visi_v2['PIRM_REG_METAI_LT'].astype('int64')
df_visi_v2['PIRM_REG_MEN_LT'] = df_visi_v2['PIRM_REG_MEN_LT'].astype('int64')
df_visi_v2['KIEKIS'] = df_visi_v2['KIEKIS'].astype('int64')
df_visi_v2

# Trinam nereikalingus stulpelius
df_visi_v2 = df_visi_v2.drop(columns=['PIRM_REG_DATA_LT_1', 'PIRM_REG_DATA_LT'])
df_visi_v2

# Įrašome į failą
df_visi_v2.to_csv('data_visi_v2.csv', index=False)

### Pagrindinis darbinis failas: data_visi_v2.csv
