==> 07_main_visi_v3.ipynb <==
## Duomenų failo versija su  4 stulpeliais

# Variantas duomenų su keturiais stulpelias
# Po bandymų su Random Forest buvo atsisakyta dėl kompiuterinių resursų trukumo
import numpy as np
import pandas as pd
df = pd.read_csv('Atviri_TP_parko_duomenys.csv', usecols=['PIRM_REG_DATA_LT', 'MARKE']) # nrows=2000,
df = df.dropna()
df = df.groupby('PIRM_REG_DATA_LT').count()['MARKE']
df = df.loc['2014-06-01':'2024-06-30']
df = df.reset_index()
df['DATA'] = pd.to_datetime(df['PIRM_REG_DATA_LT'])
df['PIRM_REG_METAI_LT'] = df['DATA'].dt.year.astype('int64')
df['PIRM_REG_MEN_LT'] = df['DATA'].dt.month.astype('int64')
df['PIRM_REG_DIENA_LT'] = df['DATA'].dt.day.astype('int64')
df['MARKE'] = df['MARKE'].astype('int64')
df = df.rename(columns = {'MARKE': 'KIEKIS'})
df = df.drop(columns=['PIRM_REG_DATA_LT', 'DATA'])
df.to_csv('data_visi_v3.csv', index=False)
df
