==> 17_sasaja.ipynb <==
## Vartotojo sąsaja

# Bibliotekų importavimas
import tkinter as tk
from tkinter import ttk
import pandas as pd
import numpy as np
import joblib
from IPython.display import display, clear_output
import matplotlib.pyplot as plt
from sklearn.metrics import mean_squared_error, r2_score
from sklearn.linear_model import Ridge
from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score
from sklearn.preprocessing import StandardScaler, PolynomialFeatures


# FUNKCIJOS
# Funkcija prognozei gauti pagal vartotojo įvestis
def predict_kiekis(reg_year, reg_men):
    # Įkeliame PolynomialFeatures ir StandardScaler
    
    # # variantas 1 Ridge
    poly = joblib.load('poly.pkl')
    scaler = joblib.load('scaler.pkl')
    loaded_model = joblib.load('regression_model.pkl')

    # # # Geriausių parametrų variantas
    # poly = joblib.load('knn_poly.pkl')
    # scaler = joblib.load('knn_scaler.pkl')
    # loaded_model = joblib.load('knn_model.pkl')

    # Paruošiame duomenis
    user_data = pd.DataFrame({
        'PIRM_REG_METAI_LT': [reg_year],
        'PIRM_REG_MEN_LT': [reg_men]
    })

    # Pritaikom polynomial savybes ir normalizavimą
    user_data_poly = poly.transform(user_data)
    user_data_scaled = scaler.transform(user_data_poly)

    # Prognozuojam pagal užkrautą modelį
    predicted_vol = loaded_model.predict(user_data_scaled)
    return predicted_vol[0]

# Funkcija duomenų įvedimui valdyti
def gauti_prognoze():
    pasirink_metai = metu_combobox.get()
    pasirink_menuo = menuo_combobox.get()
    if pasirink_metai and pasirink_menuo:
        clear_output()
        kiekis = predict_kiekis(pasirink_metai, pasirink_menuo)
        forecast_label.config(text=f'Automobilių registracijų kiekio prognozė\nData: {pasirink_metai}-{pasirink_menuo}\nKiekis: {kiekis:.0f} vnt ', background='lightblue', font='courier 15 bold')
       
    else:
        forecast_label.config('pasirinkite metus')


# MAIN 
# sukuriame pagrindinį langą
root = tk.Tk()
root.title('Automobilių registracijų kiekio prognozė')
root.geometry('600x400')
root.configure(bg='lightblue')

# Metų pasirinkimo langelis
metai = [str(metai) for metai in range(2024, 2035)]
metu_label = ttk.Label(root, text='pasirinkite metus:', background='lightblue', font='courier 20 bold')
metu_label.pack(pady=5)
metu_combobox = ttk.Combobox(root, values=metai, state='readonly')
metu_combobox.pack(pady=5)
metu_combobox.set(2024)

# Mėnesio pasirinkimo langelis
menuo = [str(menuo) for menuo in range(1, 13)]
menuo_label = ttk.Label(root, text='pasirinkite mėnesį:', background='lightblue', font='courier 20 bold')
menuo_label.pack(pady=5)
menuo_combobox = ttk.Combobox(root, values=menuo, state='readonly')
menuo_combobox.pack(pady=5)
menuo_combobox.set(1)

# Prognozės ir darbo baigimo mygtukai
prognoze_mygtukas = ttk.Button(root, text='gauti prognoze', command=gauti_prognoze, width=50)
prognoze_mygtukas.pack(pady=10)
baigti_mygtukas = ttk.Button(root, text='baigti darba', command=root.destroy, width=50)
baigti_mygtukas.pack(pady=0)

# Informacijos (rezultatų) išvedimo vietos savybės
forecast_label = ttk.Label(root, text='', background='lightblue')
forecast_label.pack(pady=10)

# Lango veikimo ciklas
root.mainloop()
